fixed second solution

+++ src/main/java/core/basesyntax/service/FileWriter.java
@@ -3,5 +3,5 @@ package core.basesyntax.service;
 import java.util.Map;
 
 public interface FileWriter {
-    void createReportFile(Map<String, Long> stockBalance, String filePath);
+    void createReportFile(Map<String, Integer> stockBalance, String filePath);
 }
